feat: utility for handling open command/context
